+++ .eslintrc.js
@@ -16,6 +16,10 @@ module.exports = {
     "react"
   ],
   "rules": {
+    "comma-dangle": [
+      "error",
+      "always"
+    ],
     "indent": [
       "error",
       2