+++ src/components/index.js
@@ -1,0 +1,5 @@
+import Email from './Email'
+import Submit from './Submit'
+import TinyLetter from './TinyLetter'
+
+export { Email, Submit, TinyLetter, }